VR-175: moving files and functions around.
