+++ src/controllers/queries/__tests__/helpers.ts
@@ -7,11 +7,10 @@ import Database from '../../../models/db/index.js'
 import { ConnectionRow, QueryRow } from '../../../models/db/types.js'
 import { UUID } from '../../../models/strings.js'
 import VeritableCloudagent from '../../../models/veritableCloudagent.js'
+import Scope3CarbonConsumptionResponseTemplates, { Scope3FormProps } from '../../../views/queries/co2scope3.js'
+import Scope3CarbonConsumptionTemplates from '../../../views/queries/newCo2scope3.js'
 import QueriesTemplates from '../../../views/queries/queries.js'
 import QueryListTemplates from '../../../views/queries/queriesList.js'
-import Scope3CarbonConsumptionResponseTemplates, { Scope3FormProps } from '../../../views/queries/responses/scope3.js'
-import Scope3CarbonConsumptionViewResponseTemplates from '../../../views/queries/responses/viewResponseToScope3.js'
-import Scope3CarbonConsumptionTemplates from '../../../views/queryTypes/scope3.js'
 
 type QueryStatus = 'resolved' | 'pending_your_input' | 'pending_their_input'
 
@@ -93,12 +92,10 @@ export const withQueriesMocks = (testOptions: Partial<QueryMockOptions> = {}) =>
   const scope3CarbonConsumptionTemplateMock = {
     newScope3CarbonConsumptionFormPage: (props: { formStage: string }) => templateListFake('scope3', props.formStage),
   } as unknown as Scope3CarbonConsumptionTemplates
-  const scope3CarbonConsumptionViewResponseTemplates = {
-    scope3CarbonConsumptionViewResponsePage: () => templateListFake('scope3Response'),
-  } as unknown as Scope3CarbonConsumptionViewResponseTemplates
 
   const scope3CarbonConsumptionResponseTemplateMock = {
     newScope3CarbonConsumptionResponseFormPage: (props: Scope3FormProps) => templateFake('queriesResponse', props),
+    view: () => templateListFake('scope3Response'),
   } as unknown as Scope3CarbonConsumptionResponseTemplates
   const queryTemplateMock = {
     chooseQueryPage: () => templateFake('queries'),
@@ -122,7 +119,7 @@ export const withQueriesMocks = (testOptions: Partial<QueryMockOptions> = {}) =>
         id: 123,
         created_at: new Date(),
         updated_at: new Date(),
-        connection_id: '11',
+        connection_id: 'aa000000-0000-0000-0000-aabbccddee00',
         query_type: 'Scope 3',
         details: 'some details',
       },
@@ -132,7 +129,6 @@ export const withQueriesMocks = (testOptions: Partial<QueryMockOptions> = {}) =>
   return {
     scope3CarbonConsumptionTemplateMock,
     scope3CarbonConsumptionResponseTemplateMock,
-    scope3CarbonConsumptionViewResponseTemplates,
     queryListTemplateMock,
     queryTemplateMock,
     mockLogger,
@@ -141,7 +137,6 @@ export const withQueriesMocks = (testOptions: Partial<QueryMockOptions> = {}) =>
     args: [
       scope3CarbonConsumptionTemplateMock,
       scope3CarbonConsumptionResponseTemplateMock,
-      scope3CarbonConsumptionViewResponseTemplates,
       queryTemplateMock,
       queryListTemplateMock,
       cloudagentMock as unknown as VeritableCloudagent,
