+++ src/views/queries/__tests__/partial-query.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { describe } from 'mocha'
-import { ConnectionRow } from '../../../../models/db/types.js'
-import Scope3CarbonConsumptionResponseTemplates from '../scope3.js'
+import { ConnectionRow } from '../../../models/db/types.js'
+import Scope3CarbonConsumptionResponseTemplates from '../co2scope3.js'
 
 const sampleDate = new Date(Date.UTC(2024, 6, 4))
 const connectionsExample: ConnectionRow[] = [
@@ -27,7 +27,7 @@ describe('Partial Query', () => {
         company_name: 'VER123',
         company_number: '3456789',
         status: 'verified_both',
-        id: '11',
+        id: 'aa000000-0000-0000-0000-aabbccddee00',
         agent_connection_id: 'agentId',
         pin_tries_remaining_count: null,
         pin_attempt_count: 0,
