+++ src/views/queries/co2scope3.tsx
@@ -1,8 +1,22 @@
 import Html from '@kitajs/html'
 import { UUID } from 'crypto'
 import { singleton } from 'tsyringe'
-import { ConnectionRow } from '../../../models/db/types.js'
-import { FormButton, LinkButton, Page } from '../../common.js'
+import { ConnectionRow } from '../../models/db/types.js'
+import { FormButton, LinkButton, Page } from '../common.js'
+
+type QueryStatus = 'resolved' | 'pending_your_input' | 'pending_their_input' | 'errored'
+
+interface Query {
+  id: string
+  company_name: string
+  query_type: string
+  updated_at: Date
+  status: QueryStatus
+  role: 'responder' | 'requester'
+  quantity: string
+  productId: string
+  emissions: string
+}
 
 export type Scope3FormStage = 'form' | 'success' | 'error'
 export interface Scope3FormProps {
@@ -162,6 +176,75 @@ export default class Scope3CarbonConsumptionResponseTemplates {
       </div>
     )
   }
+
+  public view = (query: Query): JSX.Element => (
+    <Page
+      title="Veritable - Query Response"
+      activePage="queries"
+      heading="View response to your query"
+      headerLinks={[
+        { name: 'Queries', url: '/queries' },
+        { name: query.company_name, url: `/queries/scope-3-carbon-consumption/${query.id}/view-response` },
+      ]}
+    >
+      <div class="connections header"></div>
+      <div class="card-body">
+        <div class="container-scope3-carbon">
+          <div class="box1">
+            <h1>Scope 3 Carbon Consumption</h1>
+            <p class="query-text-carbon3-consumption">
+              A query for calculationg the total scope 3 carbon consumption for a given product or component.
+            </p>
+          </div>
+          <div class="box2">
+            <h2>Query Information</h2>
+            <div class="box3 ">
+              <table>
+                <tr>
+                  <td>Supplier:</td>
+                  <td class="query-results-left-padding-table">{Html.escapeHtml(query.company_name)}</td>
+                </tr>
+                <tr>
+                  <td>ProductID:</td>
+                  <td class="query-results-left-padding-table">{Html.escapeHtml(query.productId)}</td>
+                </tr>
+                <tr>
+                  <td>Quantity:</td>
+                  <td class="query-results-left-padding-table">{Html.escapeHtml(query.quantity)}</td>
+                </tr>
+                <tr>
+                  <td>Query:</td>
+                  <td class="query-results-left-padding-table">
+                    Please provide details on the Scope 3 carbon emissions associated with the product.
+                  </td>
+                </tr>
+              </table>
+            </div>
+            <h2>Response Information</h2>
+            <div class="box3 ">
+              <div class="max-height-table">
+                <table>
+                  <tr>
+                    <td>Date Certified:</td>
+                    <td>
+                      <time class="query-results-left-padding-table">
+                        {Html.escapeHtml(new Date(query.updated_at).toISOString())}
+                      </time>
+                    </td>
+                  </tr>
+                  <tr>
+                    <td>Carbon Emissions:</td>
+                    <td class="query-results-left-padding-table">{Html.escapeHtml(query.emissions)}</td>
+                  </tr>
+                </table>
+              </div>
+            </div>
+            <LinkButton text="Back to Queries" href="/queries" style="filled" />
+          </div>
+        </div>
+      </div>
+    </Page>
+  )
 
   public tableRow = ({
     checked = false,
